+++ NEWS.md
@@ -6,6 +6,7 @@
 - Add R-Datalab Dockerfile example
 - Let Rstudio users be added with staff rights so they can install packages etc.
 - Add ability to specify disk size when creating a VM (#38) - thanks @jburos
+- Add firewall functions (#34)
 
 # googleComputeEngineR 0.1.0